streamlit_app/requirements.txt
streamlit
firecrawl-py
